+++ icx/src/main.rs
@@ -18,7 +18,7 @@ use ic_agent::{
 };
 use ic_utils::interfaces::management_canister::{
     builders::{CanisterInstall, CanisterSettings},
-    MgmtMethod,
+    BitcoinNetwork, MgmtMethod,
 };
 use ring::signature::Ed25519KeyPair;
 use std::{
@@ -290,7 +290,11 @@ pub fn get_effective_canister_id(
             | MgmtMethod::DeleteCanister
             | MgmtMethod::DepositCycles
             | MgmtMethod::UninstallCode
-            | MgmtMethod::ProvisionalTopUpCanister => {
+            | MgmtMethod::ProvisionalTopUpCanister
+            | MgmtMethod::UploadChunk
+            | MgmtMethod::ClearChunkStore
+            | MgmtMethod::StoredChunks
+            | MgmtMethod::FetchCanisterLogs => {
                 #[derive(CandidType, Deserialize)]
                 struct In {
                     canister_id: Principal,
@@ -309,6 +313,33 @@ pub fn get_effective_canister_id(
                 let in_args =
                     Decode!(arg_value, In).context("Argument is not valid for UpdateSettings")?;
                 Ok(in_args.canister_id)
+            }
+            MgmtMethod::InstallChunkedCode => {
+                #[derive(CandidType, Deserialize)]
+                struct In {
+                    target_canister: Principal,
+                }
+                let in_args = Decode!(arg_value, In)
+                    .context("Argument is not valid for InstallChunkedCode")?;
+                Ok(in_args.target_canister)
+            }
+            MgmtMethod::BitcoinGetBalanceQuery | MgmtMethod::BitcoinGetUtxosQuery => {
+                #[derive(CandidType, Deserialize)]
+                struct In {
+                    network: BitcoinNetwork,
+                }
+                let in_args = Decode!(arg_value, In)
+                    .with_context(|| format!("Argument is not valid for {method_name}"))?;
+                Ok(in_args.network.effective_canister_id())
+            }
+            MgmtMethod::BitcoinGetBalance
+            | MgmtMethod::BitcoinGetUtxos
+            | MgmtMethod::BitcoinSendTransaction
+            | MgmtMethod::BitcoinGetCurrentFeePercentiles
+            | MgmtMethod::EcdsaPublicKey
+            | MgmtMethod::SignWithEcdsa
+            | MgmtMethod::NodeMetricsHistory => {
+                bail!("Management canister method {method_name} can only be run from canisters");
             }
         }
     } else {